Refactorizar codigo.
