+++ tests/Tag/ATest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Forge\Html\Tests\Tag\Element;
+namespace Forge\Html\Tests\Tag;
 
-use Forge\Html\Tag\Element\A;
+use Forge\Html\Tag\Tag;
 use Forge\TestUtils\Assert;
 use PHPUnit\Framework\TestCase;
 
@@ -34,8 +34,8 @@ final class ATest extends TestCase
      */
     public function testCreate(array $attributes, string $content, string $expected): void
     {
-        $a = new A();
         $assert = new Assert();
-        $assert->equalsWithoutLE($expected, $a->create($attributes, $content));
+
+        $assert->equalsWithoutLE($expected, Tag::a($attributes, $content));
     }
 }
